+++ src/jax_fdm/optimization/optimizers/optimizers.py
@@ -1,6 +1,8 @@
 """
 A collection of scipy-powered, gradient-based optimizers.
 """
+import numpy as np
+
 from jax_fdm.optimization.optimizers import Optimizer
 from jax_fdm.optimization.optimizers import ConstrainedOptimizer
 from jax_fdm.optimization.optimizers import SecondOrderOptimizer
@@ -20,10 +22,28 @@ class SLSQP(ConstrainedOptimizer):
 
 class LBFGSB(Optimizer):
     """
-    The limited-memory Boyd-Fletcher-Floyd-Shannon-Byrd optimizer.
+    The limited-memory Boyd-Fletcher-Floyd-Shannon-Byrd (LBFGSB) optimizer.
     """
     def __init__(self, **kwargs):
         super().__init__(name="L-BFGS-B", disp=0, **kwargs)
+
+
+class LBFGSBS(Optimizer):
+    """
+    The limited-memory Boyd-Fletcher-Floyd-Shannon-Byrd (LBFGSB) optimizer.
+
+    This version of LBFGSB ensures compatibility of JAX gradients with scipy.
+    However, it is slower than standard LBFGSB.
+    """
+    def __init__(self, **kwargs):
+        super().__init__(name="L-BFGS-B", disp=0, **kwargs)
+
+    def gradient(self, loss):
+        """
+        Compute the gradient function of a loss function.
+        """
+        jit_gfunc = super().gradient(loss)
+        return lambda x: np.array(jit_gfunc(x))
 
 
 class BFGS(Optimizer):